% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidian.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{An Integer.}

\item{b}{A Integer}
}
\value{
The GCD of input params.
}
\description{
Implementation of the substraction based euclidean algorithm to find the Greatest Common Divisior (GCD) of two Integer numbers.
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
